% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daft_function.R
\name{daft_lyric}
\alias{daft_lyric}
\title{Lyrics from a good song}
\usage{
daft_lyric(line = NA)
}
\arguments{
\item{line}{which line? defaults to random}
}
\value{
character
}
\description{
Lyrics from a good song
}
\examples{
daft_lyric()
}
